Completes ex. 4.4, adds and tests blogs likes sum
This commit adds a function to the list_helper utility module that
calculates the total number of likes across a blog list. It takes as
a parameter an array of blog objects and returns the sum of the likes
across all blogs in the list. This commit also adds a jest unit
testing script to check that the totalLikes function behaves
as expected for input of zero blogs, one blog, and many blogs.

This commit also adds a set of test blog data in the file
tests/blogListTestData.js. It further deletes the dummy function in
the list_helper module as it is not useful, and deletes the test
script for that function.