Completes ex. 4.5, adds most liked blog util func
This commit adds the function favoriteBlog to the list_helper utility
module. This takes a list of blogs and returns an object with the
title, author, and total likes of the blog with the most likes. In a
tie it returns the last blog in the list. In an empty list, it
returns an object with blank strings for title and author, and 0 for
likes.

This also adds a unit test script for the favoriteBlog function which
tests for an empty list, a list of only 1 blog, a list of several
blogs, and a list with multiple blogs with a tie for maximum likes.