Completes ex. 4.6, adds most blogs util function
This commit adds a utility function mostBlogs to the list_helper
module. The function takes an array of blogs, and returns an object
that lists the author with the most blogs and the number of blogs
written. If the array is empty it returns an object with a blank
string as author and 0 for number of blogs. If there is a tie it
returns one of the two authors with their number of blogs written.

This commit also adds a utility module arrayFunctions.js which has
two functions, sortArray and reduceArray. The sortArray function
takes an array as input and returns it sorted using the default >, <
comparison operators. The reduceArray function takes an array and
returns an array that removes any identical adjacent members (i.e.,
if the array has 3 identical members adjacent to each other, the
returned array will have only 1, with the later non-removed array
members shifted down-index). These are used in combination by the
mostBlogs function to obtain a list of unique authors in the blog
array parameter in an efficient manner (O(log2 n)).

This commit also adds unit testing for the mostBlogs function to the
favoriteBlog.test.js jest script.