Completes ex. 4.7; adds most likes util function
This commit adds a mostLikes function to the list_helper utility
module. This function takes an array of blogs and returns an object
that consists of the author with the most total likes and the number
of total likes. If the function is called with an empty array, it
returns an object with a blank string for author and 0 for likes. If
multiple authors are tied for most likes, it returns an object with
one of their names as author and the tied number of likes as likes.

This commit also adds unit tests for the new function to the
favoriteBlog.test.js jest script.