Adds libraries and global teardwon
This commit adds dependencies to the json configuration files for the
cross-env, supertest, and express-async-errors libraries.

For cross-env, it also modifies npm run commands in package.json to
define the environment, and modifies utils/config.js to set the DB
URL depending on whether or not the process was started in a 'test'
environment.

For express-async-errors, the commit also modifies app.js to require
the library.

This commit also adds tests/teardown.js as a global teardown for
jest, and defines it as such in package.json.