Updates util modules for new libraries
This commit updates some utility modules to take advantage of the
libraries added in the previous commit. It updates logger.js not to
log when in the test environment (to avoid clutter getting in the way
of jest output and debugging logs). It also adds middleware.js, which
currently only includes an error handling function. Use of this error
handler is added into app.js.