Completes ex. 4.8, GET test & use async/await
This commit adds a new test file for database requests. It currently
includes only a test of GET for the entire database, checking that
the response is formatted as JSON and that it returns the correct
number of entries.

This commit also refactors the GET request defined in blogs.js to
use async/await instead of promises.