Completes ex. 4.9, adds and tests DB/JSON handler
This commit adds a JSON transform to the blogSchema defined in
modes/blog.js, which translates the MongoDB '_id' property into a
string property in the blog object with the name 'id' (no leading
underscore), and deletes the _id and __v properties from the JSON
data retrieved from the database.

This commit also adds a test to databaseRequests.test.js to confirm
that there is a defined property named 'id' in the blog object
retrieved from the database.