Completes ex. 4.10, POST test & use async/await
This commit modifies the POST handler in controllers/blogs.js to use
async/await instead of ordinary promise format.

This commit also adds a test to tests/databaseRequests.test.js to
validate that the POST functionality adds a correctly formatted blog
to the database. This is aided by the addition of a test/test_helper
module, which currently just includes one function which reads and
returns the contents of the database when it is called.