Completes ex. 4.11, adds&tests default likes val
This commit modifies the POST function in controllers/blogs.js so
that if the blog object sent does not contain a likes value the data
sent to the database contains a default likes value of 0. It does the
same for author with a default value of 'unknown', but this is not
a required part of the exercise.

The commit also adds a test to test/databaseRequests.js to validate
the default value for likes. It does not contain a test for the
default author value, as the exercise specifies "Do not test the
other properties of the created blogs yet."