Completes ex. 4.12, requires title&url for blog
This commit modifies the blogSchema defined in models/blog.js so that
the title and url fields are required. It also adds two tests to
test/databaseRequests.js that tests POSTing a blog object without
each of these fields and makes sure the POST returns status code 400
and that the size of the database does not change.