Completes ex. 4.15, adds users collection to DB
This commit adds functionality for a users collection to the bloglist
DB. The User model is defined in models/user.js, and contains
username, name, and passwordHash fields. Access to the users
collection is managed in controllers/users.js, and currently consists
of POST to create new users and GET to get the entire list of users.
The GET returns data with name, username, and ID formatted as a
string.

The commit also updates app.js to use the functions in
controllers/users.js.

The commit uses bcrypt to perform the password hashing - it is added
as a dependency in the package json files.