Completes ex 4.16, adds user/PW reqs to user list
This commit adds requirements to the username and password fields
when adding a new user to the users collection in the database.

For username, these are that a username be present, at least 3
characters, and unique. These are defines in the User model in
models/user.js. For uniqueness, mongoose-unique-validator is added to
the project, and set as a dependency in the package json files.

For password, the requirements are that the password be present and
be at least 3 characters. Because passwords are stored as a hash,
these requirements are defined in the POST function in
controllers/users.js.

Testing of this functionality is also added to
tests/databaseRequests.test.js. For these tests, an initial user set
is defined in tests/blogListTestData.js. Because this file is now
exporting multiple items, other test files that rely on it are
updated. There is also a new function usersInDb that is added to
tests/test_helper.js that returns the entirety of the test user
collection.