Completes ex. 4,17, links users and blogs in DB
This commit adds a field to each blog for a user to identify the user
that created the bloglist entry. It also creates an array for each
user of all the blogs they have created. The GET routines for both
display DB info for the user or blogs referenced in the blog or user
entries, respectively.

This commit also updates the POST and PUT routines for blogs to add
the first user in the database as the author. It also updates the PUT
routine so that the user entry for the first user so that if it does
not have the updated blog associated it adds that blog to the user's
list of blogs. These are temporary, as future exercises add token
-based authentication, which will be used for POST and PUT routines.

This commit also updates the blog DELETE method to remove the blog
from the user's list of blogs.