Completes ex 4.20, token extraction to middleware
This commit moves token extraction to middleware, adding a function
tokenExtractor which, if an authorization field is present that
begins with 'Bearer ', adds a token field to the request with the
encoded token string. This middleware is added before the routes in
app.js, and the request.token property is used to access the token in
the POST process for blogs.