Completes ex 4.21, requires token for DELETE & PUT
This commit requires a token for DELETE and PUT requests for the blog
collection, and further requires that the token be associated with
the user who created those entries. This is done by finding the user
associated with the token, and then checking that the request's URL
id is an entry in the blogs array in that user's entry in the
database. The check is done on the user collection side rather than
on the blog collection side to reduce await calls to the database.