Completes ex. 4.23, updates tests for token auth
This commit makes necessary updates to the JEST test file
tests/database_requests.test.js to account for token-based
authentication required for POST, DELETE, and PUT requests to blogs,
as well as making changes so that the test data now includes links
between blogs and users. It also adds tests as appropriate to test
the token-based authentication.

To assist with these tests, this commit also updates
tests/test_helper.js to add a utility function getToken which takes
as a parameter an object that includes a username and password, and
returns a string with a token for that username that can be used with
DB calls.

This commit also makes updates to controllers/blogs.js to ensure that
a correct response is sent when a PUT or DELETE request is sent to a
malformatted ID or an ID not present in the database - previously
these would return token-based errors regarding the incorrect user.
While it is technically correct that the user was not associated with
the ID submitted if the ID was malformatted or not present, it is not
a particularly informative or helpful response.

This commit also adds unknown endpoint middleware.